early-debug.el:
Note that this is not related to early-package-init.el but is an
unrelated utility to track commands and may be useful to trace events
in emacs.

code is of dubious quality. should probably be called track-events.el
as it prints a history of certain commands to a *track-events* buffer.
uses obsolete defadvice in ways that try to break it: like breaking
eager macroexpansion.  to avoid most problems the file can to be
compiled before being loaded